+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project are documented in this file.
 
+## [Unreleased]
+
+### Changed
+
+- GM mode now uses tool calling to modify game state.
+- Reworded GM system prompt to support new changes (Make sure to update or reset your custom prompts if you use them.)
+
 ## [v0.5.1] - 2025-11-10
 
 ### Fixed